requirements.txt
streamlit
langchain
cohere
requests
beautifulsoup4
pillow
langchain_community
langchain_experimental